pyproject.toml
[project]
name = "mcp-cli"
version = "0.1.0"
description = "a cli for the Model Context Provider"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "anyio>=4.6.2.post1",
    "asyncio>=3.4.3",
    "huggingface-hub>=0.26.3",
    "ollama>=0.4.2",
    "openai>=1.55.3",
    "python-dotenv>=1.0.1",
    "requests>=2.32.3",
]
